floats to doubles

+++ tests/tree_kernel_tests.cpp
@@ -18,31 +18,31 @@ struct SimpleTree{
 };
 
 BOOST_FIXTURE_TEST_CASE(kdiag1, SimpleTree){
-  vector<float> lambda = {1.0};
+  vector<double> lambda = {1.0};
   SASSTK kernel = SASSTK(lambda, false);
   kernel.Kdiag(trees, result);
   BOOST_CHECK_EQUAL(result[0].k, 6);
 };
 
 BOOST_FIXTURE_TEST_CASE(kdiag2, SimpleTree){
-  vector<float> lambda = {1.0, 0.5};
-  vector<float> alpha = {0.0, 1.0};
+  vector<double> lambda = {1.0, 0.5};
+  vector<double> alpha = {0.0, 1.0};
   SASSTK kernel = SASSTK(lambda, alpha, false);
   kernel.Kdiag(trees, result);
   BOOST_CHECK_EQUAL(result[0].k, 3);
 }
 
 BOOST_FIXTURE_TEST_CASE(kdiag3, SimpleTree){
-  vector<float> lambda = {1.0, 0.5};
-  vector<float> alpha = {0.2, 1.0};
+  vector<double> lambda = {1.0, 0.5};
+  vector<double> alpha = {0.2, 1.0};
   SASSTK kernel = SASSTK(lambda, alpha, false);
   kernel.Kdiag(trees, result);
   BOOST_CHECK_EQUAL(result[0].k, 3.44);
 }
 
 BOOST_FIXTURE_TEST_CASE(kdiag4, SimpleTree){
-  vector<float> lambda = {0.6, 0.5};
-  vector<float> alpha = {1.0, 1.0};
+  vector<double> lambda = {0.6, 0.5};
+  vector<double> alpha = {1.0, 1.0};
   SASSTK kernel = SASSTK(lambda, alpha, false);
   kernel.Kdiag(trees, result);
   BOOST_CHECK_EQUAL(result[0].k, 2.736);
